+++ src/fmpz_mod_mpoly_factor/interp.c
@@ -122,7 +122,7 @@ int fmpz_mod_polyu1n_intp_crt_sm_poly(
             FLINT_ASSERT(!fmpz_is_zero(Acoeffs + Ai));
         }
 
-        if (Fi < Flen && Ai >= 0 && Fexps[Fi] == Ai)
+        if (Fi < Flen && Ai >= 0 && Fexps[Fi] == (ulong) Ai)
         {
             /* F term ok, A term ok */
             fmpz_mod_poly_evaluate_fmpz(v, Fcoeffs + Fi, alpha, ctx);
@@ -136,7 +136,7 @@ int fmpz_mod_polyu1n_intp_crt_sm_poly(
                 Ai--;
             } while (Ai >= 0 && fmpz_is_zero(Acoeffs + Ai));
         }
-        else if (Fi < Flen && (Ai < 0 || Fexps[Fi] > Ai))
+        else if (Fi < Flen && (Ai < 0 || Fexps[Fi] > (ulong) Ai))
         {
             /* F term ok, A term missing */
             fmpz_mod_poly_evaluate_fmpz(v, Fcoeffs + Fi, alpha, ctx);
@@ -590,7 +590,7 @@ int fmpz_mod_mpolyn_interp_crt_sm_poly(
 
         mpoly_monomial_zero(Texp + N*Ti, N);
 
-        if (Fi < Flen && Ai >= 0 && ((Fexp + N*Fi)[off]>>shift) == Ai)
+        if (Fi < Flen && Ai >= 0 && ((Fexp + N*Fi)[off]>>shift) == (ulong) Ai)
         {
             /* F term ok, A term ok */
             fmpz_mod_poly_evaluate_fmpz(u, Fcoeff + Fi, alpha, ctx->ffinfo);
@@ -612,7 +612,7 @@ int fmpz_mod_mpolyn_interp_crt_sm_poly(
                 Ai--;
             } while (Ai >= 0 && fmpz_is_zero(Acoeff + Ai));
         }
-        else if (Fi < Flen && (Ai < 0 || ((Fexp + N*Fi)[off]>>shift) > Ai))
+        else if (Fi < Flen && (Ai < 0 || ((Fexp + N*Fi)[off]>>shift) > (ulong) Ai))
         {
             /* F term ok, A term missing */
             fmpz_mod_poly_evaluate_fmpz(v, Fcoeff + Fi, alpha, ctx->ffinfo);
@@ -630,7 +630,7 @@ int fmpz_mod_mpolyn_interp_crt_sm_poly(
             (Texp + N*Ti)[off] = (Fexp + N*Fi)[off];
             Fi++;
         }
-        else if (Ai >= 0 && (Fi >= Flen || ((Fexp + N*Fi)[off]>>shift) < Ai))
+        else if (Ai >= 0 && (Fi >= Flen || ((Fexp + N*Fi)[off]>>shift) < (ulong) Ai))
         {
             /* F term missing, A term ok */
             changed = 1;
